all fix + add custom exception

+++ src/main/resources/init_db.sql
@@ -1,7 +1,7 @@
 create schema if not exists test;
 
 create table if not exists book (
-    book_id serial primary key,
+    book_id bigint primary key,
     title varchar(20),
     price decimal
 );
